serverless/destroy-resources/requirements.txt
google-cloud-storage
python-dotenv
functions-framework
pyyaml
google-cloud-functions
firebase-admin